# flask_app/core/swagger/swagger.json
{"info": {"description": "Auth API", "title": "Movies Auth Service", "version": "0.0.1"}, "servers": [{"description": "localhost", "url": "http://127.0.0.1"}], "tags": [{"name": "Authentification", "description": "Endpoints related to Authentification"}, {"name": "Tokens", "description": "Token-related routes"}, {"name": "Roles", "description": "Roles routes"}, {"name": "Additional", "description": "Other routes"}], "paths": {"/auth/authorize": {"get": {"summary": "Check if user is logged in", "description": "Returns main user information for success\n", "tags": ["Authentification"], "parameters": [{"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/logout": {"post": {"summary": "Loggs out user", "description": "deletes JWT cookies and places them in black list\n", "tags": ["Authentification"], "parameters": [{"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/sign-in": {"post": {"summary": "Sign-ins a user", "description": "In success returns pair of JWT\n", "tags": ["Authentification"], "parameters": [{"name": "login", "in": "json", "example": "your@mail.com", "description": "User email", "required": true, "schema": {"type": "string"}}, {"name": "password", "in": "json", "example": "123qwer", "description": "User password", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/sign-up": {"post": {"summary": "Registers user", "description": "In success returns pair of JWT\n", "tags": ["Authentification"], "parameters": [{"name": "login", "in": "json", "example": "your@mail.com", "description": "User email", "required": true, "schema": {"type": "string"}}, {"name": "password", "in": "json", "example": "123qwer", "description": "User password", "required": true, "schema": {"type": "string"}}, {"name": "first_name", "in": "json", "example": "Ivan", "description": "User First Name", "required": true, "schema": {"type": "string"}}, {"name": "last_name", "in": "json", "example": "Ivanov", "description": "User Last Name", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "User successfully registered"}, "401": {"description": "User already exists"}}}}, "/auth/refresh": {"post": {"summary": "Gives new pair of JWT", "description": "Removes old tokens and places them in a black list. Gives new pair of tokens.\n", "tags": ["Tokens"], "parameters": [{"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/change-role": {"post": {"summary": "Adds or deletes user role", "description": "Ruteruns updates user rolse for success\n", "tags": ["Roles"], "parameters": [{"name": "user_id", "in": "json", "required": true, "schema": {"type": "string"}}, {"name": "role", "in": "json", "required": true, "schema": {"type": "string"}}, {"name": "action", "in": "json", "required": true, "schema": {"type": "string"}}, {"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/get-user-description": {"get": {"summary": "Dives description for provided user", "tags": ["Additional"], "parameters": [{"name": "user_id", "in": "json", "required": true, "schema": {"type": "string"}}, {"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}, "/auth/sign-in-history": {"get": {"summary": "Gives login history", "tags": ["Additional"], "parameters": [{"name": "access_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}, {"name": "refresh_token_cookie", "in": "cookie", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Ok"}}}}}, "openapi": "3.0.2"}